feat: :fire: Move ShapeCreator and TextViewAdapter
