+++ app/src/main/java/com/example/jobcpp/View/Utils/TextViewAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.jobcpp.View.Components.Utils
+package com.example.jobcpp.View.Utils
 
 import android.content.Context
 import android.view.View
@@ -21,6 +21,7 @@ class TextViewAdapter(private val context: Context, private val items: List<Text
     }
 
     override fun getView(position: Int, convertView: View?, parent: ViewGroup?): View {
+
         val textView = if (convertView == null) {
             TextView(context).apply {
                 layoutParams = items[position].layoutParams
@@ -29,12 +30,10 @@ class TextViewAdapter(private val context: Context, private val items: List<Text
             convertView as TextView
         }
 
-        // Configura el TextView según sea necesario
         textView.text = items[position].text
         textView.background = items[position].background
         textView.setTextColor(items[position].currentTextColor)
         textView.gravity = items[position].gravity
-
 
         return textView
     }
